hw 03 #2 completion

+++ src/homework/c_decisions/main.py
@@ -1,7 +1,7 @@
 from decisions import get_options_ratio, get_faculty_rating
 
 def main():
-    options = int(input("Enter the number of options: "))
+    options = float(input("Enter the number of options: "))
     total = float(input("Enter the total: "))
     ratio = get_options_ratio(options, total)
     rating = get_faculty_rating(ratio)
